requirements.txt
# Core ML libraries
torch>=1.9.0
torchvision>=0.10.0
facenet-pytorch>=2.5.0

# Computer Vision
opencv-python>=4.5.0
mtcnn>=0.1.0
Pillow>=8.0.0

# Utilities
numpy>=1.19.0
scikit-learn>=1.0.0
tqdm>=4.60.0

# Web API
flask>=2.0.0
flask-cors>=3.0.0
pyngrok>=5.0.0

# Optional: TensorFlow (jika diperlukan untuk komponen lain)
# tensorflow>=2.6.0

# Plotting (untuk testing)
matplotlib>=3.3.0